fix(security): resolve 25 PHPCS security findings
- PreparedSQL: use %i identifier placeholder in LogMiddleware (real SQL injection fix)
- PreparedSQL: suppress false positive in AuditLogRepository (safe dynamic WHERE with placeholders)
- NonceVerification: suppress false positives in 3 AJAX controllers (nonce checked via verify_request())
- LikeWildcards: suppress false positive in MediaAjaxController (fixed pattern, not user input)
- EscapeOutput: add esc_html()/wp_kses_post()/(int) casts for 5 genuine findings
- EscapeOutput: suppress 8 intentionally raw outputs (SSE/JSON-LD/XML/Markdown feeds)

+++ modules/analytics/templates/dashboard.php
@@ -266,7 +266,7 @@ $has_usage_data = ( $usage_stats['total']['requests'] ?? 0 ) > 0;
             </span>
             <?php if ( $status['state'] === 'open' && $status['recovery_in'] ) : ?>
             <span class="wpmind-status-recovery">
-                <?php printf( esc_html__( '%d分钟后恢复', 'wpmind' ), ceil( $status['recovery_in'] / 60 ) ); ?>
+                <?php printf( esc_html__( '%d分钟后恢复', 'wpmind' ), (int) ceil( $status['recovery_in'] / 60 ) ); ?>
             </span>
             <?php endif; ?>
         </div>
